Feature/final

+++ src/main/java/br/com/itau/geradornotafiscal/factory/NotaFiscalFactory.java
@@ -5,6 +5,7 @@ import br.com.itau.geradornotafiscal.model.Pedido;
 import br.com.itau.geradornotafiscal.model.ItemNotaFiscal;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -12,7 +13,7 @@ import java.util.UUID;
 @Component
 public class NotaFiscalFactory {
 
-    public NotaFiscal criarNotaFiscal(Pedido pedido, List<ItemNotaFiscal> itensNotaFiscal, double valorFrete) {
+    public NotaFiscal criarNotaFiscal(Pedido pedido, List<ItemNotaFiscal> itensNotaFiscal, BigDecimal valorFrete) {
         return NotaFiscal.builder()
                 .idNotaFiscal(UUID.randomUUID().toString())
                 .data(LocalDateTime.now())
