+++ src/main/java/br/com/itau/geradornotafiscal/service/CalculadoraAliquotaService.java
@@ -2,6 +2,8 @@ package br.com.itau.geradornotafiscal.service;
 
 import br.com.itau.geradornotafiscal.model.Pedido;
 
+import java.math.BigDecimal;
+
 public interface CalculadoraAliquotaService {
-    double calcularAliquota(Pedido pedido);
+    BigDecimal calcularAliquota(Pedido pedido);
 }
